Add e2e test for check-fjord script
